+++ src/git/utils/-webview/repository.utils.ts
@@ -1,4 +1,4 @@
-import { convertRemoteProviderIdToIntegrationId } from '../../../plus/integrations/utils/-webview/integration.utils';
+import { getIntegrationIdForRemote } from '../../../plus/integrations/utils/-webview/integration.utils';
 import { configuration } from '../../../system/-webview/configuration';
 import { formatDate, fromNow } from '../../../system/date';
 import { map } from '../../../system/iterable';
@@ -78,13 +78,16 @@ export async function toRepositoryShapeWithProvider(
 			icon: remote.provider.icon === 'remote' ? 'cloud' : remote.provider.icon,
 			integration: remote.supportsIntegration()
 				? {
-						id: convertRemoteProviderIdToIntegrationId(remote.provider.id)!,
+						id: getIntegrationIdForRemote(remote.provider)!,
 						connected: remote.maybeIntegrationConnected ?? false,
 					}
 				: undefined,
 			supportedFeatures: remote.provider.supportedFeatures,
 			url: await remote.provider.url({ type: RemoteResourceType.Repo }),
 		};
+		if (provider.integration?.id == null) {
+			provider.integration = undefined;
+		}
 	}
 
 	return { ...toRepositoryShape(repo), provider: provider };